# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Oqu REST API",
        "contact": {}
    },
    "paths": {
...
        "/api/courses": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Course"
                            }
                        }
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CourseDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Provide number"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/courses/{id}/enroll": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully enrolled!"
                    },
                    "400": {
                        "description": "Incorrect user id or provide number error"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [user] can access! Your role is [admin or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/courses/{id}/reset": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "course"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Course id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course rating reset"
                    },
                    "400": {
                        "description": "Incorrect user id or provide number error"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [user] can access! Your role is [admin or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/users/enrollments": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Course"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect user id"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [user] can access! Your role is [admin or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/users/profile": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Incorrect user id"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/users/rating": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Rating"
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect user id"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [user] can access! Your role is [admin or moderator]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/moderator/comments": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderator"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ModeratorCommentView"
                            }
                        }
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [moderator] can access! Your role is [admin or user]"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/moderator/comments/{id}": {
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "moderator"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Comment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DeletedComment"
                        }
                    },
                    "400": {
                        "description": "Provide number"
                    },
                    "401": {
                        "description": "No token found, incorrect token or token expired"
                    },
                    "403": {
                        "description": "Only [moderator] can access! Your role is [admin or user]"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Course": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CourseDetails": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lessons": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.CourseLesson"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.CourseLesson": {
            "type": "object",
            "properties": {
                "completed": {
                    "type": "boolean"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "models.DeletedComment": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "models.ModeratorCommentView": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "course_id": {
                    "type": "integer"
                },
                "course_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "posted_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.Rating": {
            "type": "object",
            "properties": {
                "course_name": {
                    "type": "string"
                },
                "total_score": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}